backend/pyproject.toml
[project]
name = "elin-backend"
version = "0.1.0"
description = "FastAPI backend for ELIN 神域引擎"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2,<2.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "fastapi>=0.111.0,<1.0.0",
  "pwdlib>=0.2.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "pyjwt>=2.9.0,<3.0.0",
  "openai>=2.21.0",
  "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pre-commit>=3.8.0,<4.0.0",
  "ruff>=0.6.5,<1.0.0"
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
  ".venv",
  "__pycache__"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501", "I001", "UP007", "UP035"]
